requirements.txt
requests==2.25.1
python_telegram_bot==13.3
vk_api==11.9.3